feat: 카드 상세 조회 로직 #32
- 댓글 UI 개발
- 카드 상세 화면 이동 로직 및 쓰기 로직 구현
- 카드 상세 화면 UI 에 대한 로직 구현

+++ data/network/src/main/java/com/phew/network/retrofit/CardDetailsInquiryHttp.kt
@@ -22,7 +22,7 @@ interface CardDetailsInquiryHttp {
      */
     @POST(BuildConfig.API_URL_CARD_LIKE)
     suspend fun requestCardLike(
-        @Path("cardId") cardId: Int
+        @Path("cardId") cardId: Long
     ) : Response<Unit>
 
     /**
@@ -30,7 +30,7 @@ interface CardDetailsInquiryHttp {
      */
     @DELETE(BuildConfig.API_URL_CARD_LIKE)
     suspend fun deleteCardLike(
-        @Path("cardId") cardId: Int
+        @Path("cardId") cardId: Long
     ): Response<Unit>
 
     /**
@@ -38,7 +38,7 @@ interface CardDetailsInquiryHttp {
      */
     @GET(BuildConfig.API_URL_CARD_DETAIL)
     suspend fun requestCardDetail(
-        @Path("cardId") cardId: Int,
+        @Path("cardId") cardId: Long,
         @Query("latitude") latitude: Double? = null,
         @Query("longitude") longitude: Double? = null
     ): Response<CardDetailResponseDTO>
@@ -48,7 +48,7 @@ interface CardDetailsInquiryHttp {
      */
     @POST(BuildConfig.API_URL_CARD_DETAIL)
     suspend fun postCardDetail(
-        @Path("cardId") cardId: Int,
+        @Path("cardId") cardId: Long,
         @Body body: RequestUploadCardAnswerDTO
     ): Response<CardReplyResponseDTO>
 
@@ -57,7 +57,7 @@ interface CardDetailsInquiryHttp {
      */
     @DELETE(BuildConfig.API_URL_CARD_DELETE)
     suspend fun deleteCard(
-        @Path("cardId") cardId: Int
+        @Path("cardId") cardId: Long
     ): Response<Unit>
 
     /**
@@ -65,7 +65,7 @@ interface CardDetailsInquiryHttp {
      */
     @GET(BuildConfig.API_URL_CARD_COMMENT)
     suspend fun requestCardComments(
-        @Path("cardId") cardId: Int,
+        @Path("cardId") cardId: Long,
         @Query("latitude") latitude: Double? = null,
         @Query("longitude") longitude: Double? = null
     ): Response<List<CardCommentResponseDTO>>
@@ -75,8 +75,8 @@ interface CardDetailsInquiryHttp {
      */
     @GET(BuildConfig.API_URL_CARD_COMMENT_MORE)
     suspend fun requestCardCommentsMore(
-        @Path("cardId") cardId: Int,
-        @Path("lastId") lastId: Int,
+        @Path("cardId") cardId: Long,
+        @Path("lastId") lastId: Long,
         @Query("latitude") latitude: Double? = null,
         @Query("longitude") longitude: Double? = null
     ): Response<List<CardCommentResponseDTO>>
