+++ data/repository/src/main/java/com/phew/repository/network/MockCardFeedRepositoryImpl.kt
@@ -105,7 +105,7 @@ class MockCardFeedRepositoryImpl @Inject constructor() : CardFeedRepository {
     }
 
     override suspend fun requestUploadCardAnswer(
-        cardId: Int,
+        cardId: Long,
         isDistanceShared: Boolean,
         latitude: Double?,
         longitude: Double?,
